requirements.txt
asgiref==3.8.1
asttokens==2.4.1
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
executing==2.0.1
importlib-metadata==8.0.0
install==1.3.5
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jupyter-client==8.6.2
jupyter-core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt-toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.0.3
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.12.2
wcwidth==0.2.13
zipp==3.19.2
